fix: correct all id in each model to Integer to match with DB

+++ src/main/java/com/recipes/fromfridge/repository/FridgeIngredientRepository.java
@@ -4,7 +4,7 @@ import com.recipes.fromfridge.model.FridgeIngredient;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
-public interface FridgeIngredientRepository extends JpaRepository<FridgeIngredient, Long> {
+public interface FridgeIngredientRepository extends JpaRepository<FridgeIngredient, Integer> {
 
     // Check if an ingredient is already in fridge
     boolean existsByIngredientId(Integer ingredientId);
